requirements.txt
Flask==2.3.2
Werkzeug==2.3.7
python-dotenv==1.0.0
Flask-Marshmallow==0.14.0
Marshmallow==3.19.0
